pyproject.toml
[project]
name = "unixagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu130" },
]
torchvision = [
    { index = "pytorch-cu130" },
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
